Make script more efficient

+++ scripts/install-neovim.sh
@@ -10,10 +10,12 @@ if command -v apt-get > /dev/null; then # Install for debian-based distros
 fi
 
 # Install neovim
-if [ "$(apt-cache show neovim | grep Version | awk -F '.' '{print $2}')" -ge 9 ]; then # Install for debian-based distros
+if command -v apt-get > /dev/null && [ "$(apt-cache show neovim | grep Version | awk -F '.' '{print $2}')" -ge 9 ]; then # Install for debian-based distros
   sudo apt-get install -y neovim
 else
-  ./install-snap.sh
+  if ! command -v snap > /dev/null; then
+    ./install-snap.sh
+  fi
   sudo snap install nvim 2>/dev/null || sudo snap install nvim --classic
   sudo ln -s /snap/bin/nvim /usr/local/bin/
 fi
